Não usar tempdir no windows e não ignorar possiveis erros ao tentar deletar e criar diretorio .tempdir

+++ utils/music/local_lavalink.py
@@ -194,17 +194,14 @@ def run_lavalink(
     if 0 < lavalink_initial_ram < lavalink_ram_limit:
         java_cmd += f" -Xms{lavalink_ram_limit}m"
 
-    try:
-        shutil.rmtree("./.tempjar")
-    except:
-        pass
+    if os.name != "nt":
 
-    try:
+        if os.path.isdir("./.tempjar"):
+            shutil.rmtree("./.tempjar")
+
         os.makedirs("./.tempjar/undertow-docbase.80.2258596138812103750")
-    except:
-        pass
 
-    java_cmd += f" -Djava.io.tmpdir={os.getcwd()}/.tempjar -jar Lavalink.jar"
+        java_cmd += f" -Djava.io.tmpdir={os.getcwd()}/.tempjar -jar Lavalink.jar"
 
     if clear_plugins:
         try:
